add new load and concurrency options to reporter and printer
